requirements.txt
-r requirements/prod.txt

cumulusci~=3.84.1
jmespath~=1.0.1
py~=1.11.0
colorama~=0.4.6
pytest~=7.0.1
beautifulsoup4~=4.12.2
soupsieve~=2.5
typing_extensions~=4.7.1
pip~=24.0
attrs~=23.1.0
distro~=1.8.0
wheel~=0.41.2
Pillow~=9.4.0
Sphinx~=5.3.0
tornado~=6.3.3
filelock~=3.13.1
platformdirs~=3.10.0
tox~=4.11.4
tomli~=2.0.1
pluggy~=1.0.0
packaging~=23.2
chardet~=5.2.0
virtualenv~=20.24.7
cachetools~=5.3.2
multidict~=6.0.4
yarl~=1.9.3
wrapt~=1.16.0
setuptools~=68.2.2
isort~=5.12.0
Babel~=2.13.1
black~=23.11.0
pathspec~=0.11.2
identify~=2.5.32
flake8~=3.9.2
pyflakes~=2.3.1
snowballstemmer~=2.2.0
imagesize~=1.4.1
Deprecated~=1.2.14
iniconfig~=2.0.0
pexpect~=4.8.0
numpy~=1.23.2
Werkzeug~=3.0.1
github3.py~=4.0.1
coverage~=7.3.2
oauthlib~=3.2.2
decorator~=5.1.1
alabaster~=0.7.13
responses~=0.23.1
typeguard~=2.13.3
referencing~=0.31.1
jsonschema~=4.20.0
jsonpointer~=2.1
cfgv~=3.4.0
distlib~=0.3.7
testfixtures~=7.2.2
atlassian-python-api~=3.41.10